special message fix code smel
