+++ apps/frontend/src/components/marketplace/types.ts
@@ -1,0 +1,11 @@
+interface SpecialMessageInterface {
+    type: string;
+    data: {
+        id: string;
+        [key: string]: any;
+    };
+}
+
+export {
+    SpecialMessageInterface
+}
